model/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.1.3
fastapi==0.103.1
gotrue==1.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
joblib==1.3.2
numpy==1.26.0
packaging==23.1
postgrest==0.10.8
pydantic==2.3.0
pydantic_core==2.6.3
PyMuPDF==1.22.5
PyMuPDFb==1.23.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
realtime==1.0.0
scikit-learn==1.1.2
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
threadpoolctl==3.2.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==10.4
